requirements.txt
filestack-python==3.5.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
gunicorn==20.1.0
Jinja2==3.1.2
psycopg2==2.9.5
python-dotenv==0.21.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
WTForms==3.0.1
